fix for price representation in db, added hint for price fields

+++ db/migrate/20140316093221_add_district_price_room_to_venues.rb
@@ -1,6 +1,6 @@
 class AddDistrictPriceRoomToVenues < ActiveRecord::Migration
   def change
     add_column :venues, :district, :string
-    add_column :venues, :price_room, :integer
+    add_column :venues, :price_room, :decimal
   end
 end
